+++ tests/PhpSpreadsheetTests/Calculation/TextDataTest.php
@@ -13,6 +13,17 @@ class TextDataTest extends TestCase
     public function setUp()
     {
         Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    public function tearDown()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
     }
 
     /**
@@ -399,5 +410,22 @@ class TextDataTest extends TestCase
     public function providerTEXTJOIN()
     {
         return require 'data/Calculation/TextData/TEXTJOIN.php';
+    }
+
+    /**
+     * @dataProvider providerNUMBERVALUE
+     *
+     * @param mixed $expectedResult
+     * @param array $args
+     */
+    public function testNUMBERVALUE($expectedResult, $args)
+    {
+        $result = TextData::NUMBERVALUE(...$args);
+        self::assertEquals($expectedResult, $result);
+    }
+
+    public function providerNUMBERVALUE()
+    {
+        return require 'data/Calculation/TextData/NUMBERVALUE.php';
     }
 }